Feat: remove care and n2n configs, reduce code footprint
