+++ src/careamics/config/ng_configs/__init__.py
@@ -1,7 +1,7 @@
 """Definitions of configurations for CAREamics, compatible with the NG dataset."""
 
-__all__ = ["CAREConfiguration", "N2NConfiguration", "N2VConfiguration"]
+__all__ = ["N2VConfiguration", "NGConfiguration"]
 
-from .care_configuration import CAREConfiguration
-from .n2n_configuration import N2NConfiguration
+
 from .n2v_configuration import N2VConfiguration
+from .ng_configuration import NGConfiguration
