--- src/careamics/config/ng_configs/care_configuration.py
@@ -1,11 +1,0 @@
-"""Configuration for N2V."""
-
-from careamics.config.algorithms import CAREAlgorithm
-
-from .ng_configuration import NGConfiguration
-
-
-class CAREConfiguration(NGConfiguration):
-    """CARE-specific configuration."""
-
-    algorithm_config: CAREAlgorithm
