--- src/careamics/config/ng_configs/n2n_configuration.py
@@ -1,11 +1,0 @@
-"""Configuration for N2N."""
-
-from careamics.config.algorithms import N2NAlgorithm
-
-from .ng_configuration import NGConfiguration
-
-
-class N2NConfiguration(NGConfiguration):
-    """N2N-specific configuration."""
-
-    algorithm_config: N2NAlgorithm
