+++ src/careamics/config/ng_factories/__init__.py
@@ -2,14 +2,21 @@
 
 __all__ = [
     "create_advanced_care_config",
+    "create_advanced_n2n_config",
     "create_advanced_n2v_config",
     "create_care_config",
+    "create_n2n_config",
     "create_n2v_config",
     "create_ng_data_configuration",
     "create_structn2v_config",
 ]
 
-from .care_factory import create_advanced_care_config, create_care_config
+from .care_n2n_factory import (
+    create_advanced_care_config,
+    create_advanced_n2n_config,
+    create_care_config,
+    create_n2n_config,
+)
 from .data_factory import create_ng_data_configuration
 from .n2v_factory import (
     create_advanced_n2v_config,
