+++ tests/config/ng_factories/test_care_factory.py
@@ -1,6 +1,7 @@
 import pytest
 
-from careamics.config.ng_configs import CAREConfiguration
+from careamics.config.algorithms import CAREAlgorithm
+from careamics.config.ng_configs import NGConfiguration
 from careamics.config.ng_factories import (
     create_advanced_care_config,
     create_care_config,
@@ -19,7 +20,8 @@ class TestStandardConfig:
             patch_size=[64, 64],
             batch_size=8,
         )
-        assert isinstance(config, CAREConfiguration)
+        assert isinstance(config, NGConfiguration)
+        assert isinstance(config.algorithm_config, CAREAlgorithm)
 
     def test_no_aug(self):
         """Test no augmentation."""
@@ -135,7 +137,7 @@ class TestAdvancedCAREConfig:
         """Test that num_workers can be set and overrriden by train dataloader."""
         num_workers = 4
 
-        config: CAREConfiguration = create_advanced_care_config(
+        config: NGConfiguration = create_advanced_care_config(
             experiment_name="test",
             data_type="tiff",
             axes="YX",
@@ -149,7 +151,7 @@ class TestAdvancedCAREConfig:
 
         # test overrride
         alt_num_workers = 2
-        config: CAREConfiguration = create_advanced_care_config(
+        config: NGConfiguration = create_advanced_care_config(
             experiment_name="test",
             data_type="tiff",
             axes="YX",
